这次修改比较大,几次没提交,做的东西比较多,1修改数据库脚本,2添加了触发器,3处理入库出库库存问题, 还没改好代码,明天在修改

+++ src/main/java/org/mo/jims/coop/service/ProviderInfoService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.annotation.Resource;
 
 import org.mo.jims.coop.entity.ProviderInfo;
+import org.mo.jims.coop.entity.ProviderGood;
 import org.mo.jims.coop.repository.ProviderInfoRepository;
 import org.mo.open.common.util.Page;
 import org.springframework.stereotype.Service;
@@ -16,28 +17,48 @@ import org.springframework.transaction.annotation.Transactional;
 public class ProviderInfoService {
 
 	private ProviderInfoRepository providerInfoRepository;
-
+	
 	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
-	public ProviderInfo getProviderInfoByName(String name) {
-		ProviderInfo selcetByName = null;
-		selcetByName = providerInfoRepository.selectByName(name);
-		if (selcetByName == null) {
-			return null;
+	public List<ProviderInfo> getProviderInfoByGood(String name){
+		if (!"".equals(name) && name != null) {
+			return providerInfoRepository.selectByGood(name);
 		}
-		return selcetByName;
+		return null;
 	}
 	
 	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
-	public Page<ProviderInfo> getProviderInfoByNameOrAbbreviation(String name,
+	public List<String> getAllProviderName(){
+		return providerInfoRepository.selectAllProviderName();
+	}
+	
+	@Transactional(noRollbackFor = Exception.class)
+	public Boolean batchRemove(String[] id){
+		if (id != null) {
+			providerInfoRepository.batchDelete(id);
+			return true;
+		}
+		return false;
+	}
+	
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public ProviderInfo getProviderInfoByName(String name) {
+		if (name != null && !"".equals(name)) {
+			return providerInfoRepository.selectByName(name);
+		}
+		return null;
+	}
+	
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public Page<ProviderInfo> getProviderInfoByCriteria(String name,
 			Date beginTime, Date endTime, int page, int pageSize) {
 		Page<ProviderInfo> providerInfoPage = new Page<ProviderInfo>();
 		providerInfoPage.setTotalElement(providerInfoRepository
 				.countAllByCriteria(name, beginTime, endTime), pageSize);
+		providerInfoPage.setPageSize(pageSize);
+		providerInfoPage.setCurrentPage(page);
 		if (providerInfoPage.getTotalElement() == 0) {
 			return providerInfoPage;
 		}
-		providerInfoPage.setPageSize(pageSize);
-		providerInfoPage.setCurrentPage(page);
 		List<ProviderInfo> selectAllByNameOrAbbreviation = providerInfoRepository
 				.selectAllByCriteria(name, beginTime, endTime, (page - 1)
 						* pageSize, pageSize);
@@ -46,26 +67,90 @@ public class ProviderInfoService {
 	}
 
 	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
-	public ProviderInfo getByPK(String id) {
-		return providerInfoRepository.selectByPK(id);
+	public ProviderInfo getProviderInfoByPK(String id) {
+		if(id!=null&&!"".equals(id)){
+			return providerInfoRepository.selectByPK(id);
+		}
+		return null;
 	}
 
 	@Transactional(noRollbackFor = Exception.class)
-	public boolean save(ProviderInfo entity) {
-		providerInfoRepository.insert(entity);
-		return true;
+	public boolean saveProviderInfo(ProviderInfo entity) {
+		if (entity != null) {
+			if (entity.getTime() == null) {
+				entity.setTime(providerInfoRepository.getCurrentTime());
+			}
+			providerInfoRepository.insert(entity);
+			return true;
+		}
+		return false;
 	}
 
 	@Transactional(noRollbackFor = Exception.class)
-	public boolean alter(ProviderInfo entity) {
-		providerInfoRepository.updateByPK(entity);
-		return true;
+	public boolean alterProviderInfo(ProviderInfo entity) {
+		if (entity != null) {
+			providerInfoRepository.updateByPK(entity);
+			return true;
+		}
+		return false;
 	}
 
 	@Transactional(noRollbackFor = Exception.class)
-	public boolean removeByPK(String id) {
-		providerInfoRepository.deleteByPK(id);
-		return true;
+	public boolean removeProviderInfoByPK(String id) {
+		if(id!=null && !"".equals(id)){
+			ProviderInfo selectByPK = providerInfoRepository.selectByPK(id);
+			ProviderGood providerGood = new ProviderGood();
+			providerGood.setProviderInfo(selectByPK);
+			providerInfoRepository.deleteByPK(id);
+			providerInfoRepository.deleteRelativity(providerGood);
+			return true;
+		}
+		return false;
+	}
+	
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInToday() {
+		return providerInfoRepository.selectToday();
+	}
+
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInYesterday() {
+		return providerInfoRepository.selectYesterday();
+	}
+
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInThisWeek() {
+		return providerInfoRepository.selectThisWeek();
+	}
+
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInLastWeek() {
+		return providerInfoRepository.selectLastWeek();
+	}
+	
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInThisMonth() {
+		return providerInfoRepository.selectThisMonth();
+	}
+
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInLastMonth() {
+		return providerInfoRepository.selectLastMonth();
+	}
+	
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInThisYear() {
+		return providerInfoRepository.selectThisYear();
+	}
+	
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getProviderInfoInLastYear() {
+		return providerInfoRepository.selectLastYear();
+	}
+	
+	@Transactional(readOnly = true, propagation = Propagation.NOT_SUPPORTED)
+	public List<ProviderInfo> getAllProviderInfo(){
+		return providerInfoRepository.selectAll();
 	}
 
 	public ProviderInfoRepository getProviderInfoRepository() {
